Add a Identity basic sample
This is a sample to show how the basic authentication functionalities works with Identity Core and Entity Framework with SQLite.

The functionalities provided here are:
 * Register a user
 * Sign in
 * Sign out
 * Access to a page restricted to the authenticated users

Note: SQLite was chosen to have a standalone working sample application because SQLite is contained as a DLL in the solution but there are a lots of other database providers allowed.

Note: related to the issue #159
Note2: sorry for the delay, I was sick :/